test: update the test for type augmentation
