+++ test/types/connection-promise/tsconfig.json
@@ -1,0 +1,3 @@
+{
+  "rootDir": "."
+}
